backend/requirements.txt
fastapi==0.135.1
uvicorn[standard]==0.41.0
pydantic==2.12.5
pydantic-settings==2.13.1
python-multipart==0.0.9
requests==2.32.5
numpy==2.2.6
Pillow==12.1.1
gdown==5.2.1
sentence-transformers==2.7.0
faiss-cpu==1.13.2
qdrant-client==1.8.2
rank-bm25==0.2.2
openai==2.28.0
pytest==9.0.2
httpx==0.28.1
